Upstream Mediatek MT6795 platform code. Do more code sharing with mt8173 platform. 2nd round pull request
