+++ plat/mediatek/mt6795/drivers/uart/8250_console.S
@@ -50,69 +50,7 @@
 	 * -----------------------------------------------
 	 */
 func console_core_init
-	b core_init_fail
-	/* Check the input base address */
-	cbz	x0, core_init_fail
-	/* Check baud rate and uart clock for sanity */
-	cbz	w1, core_init_fail
-	cbz	w2, core_init_fail
-
-	/* Disable interrupt */
-	str	wzr, [x0, #UART_IER]
-
-	/* Force DTR and RTS to high */
-	mov	w3, #(UART_MCR_DTR | UART_MCR_RTS)
-	str	w3, [x0, #UART_MCR]
-
-	/* Check high speed */
-	movz	w3, #:abs_g1:115200
-	movk	w3, #:abs_g0_nc:115200
-	cmp	w2, w3
-	b.hi	1f
-
-	/* Non high speed */
-	lsl	w2, w2, #4
-	mov	w3, wzr
-	b	2f
-
-	/* High speed */
-1:	lsl	w2, w2, #2
-	mov	w3, #2
-
-	/* Set high speed UART register */
-2:	str	w3, [x0, #UART_HIGHSPEED]
-
-	/* Calculate divisor */
-	udiv	w3, w1, w2	/* divisor = uartclk / (quot * baudrate) */
-	msub	w1, w3, w2, w1	/* remainder = uartclk % (quot * baudrate) */
-	lsr	w2, w2, #1
-	cmp	w1, w2
-	cinc	w3, w3, hs
-
-	/* Set line configuration, access divisor latches */
-	mov	w1, #(UART_LCR_DLAB | UART_LCR_WLS_8)
-	str	w1, [x0, #UART_LCR]
-
-	/* Set the divisor */
-	and	w1, w3, #0xff
-	str	w1, [x0, #UART_DLL]
-	lsr	w1, w3, #8
-	and	w1, w1, #0xff
-	str	w1, [x0, #UART_DLH]
-
-	/* Hide the divisor latches */
-	mov	w1, #UART_LCR_WLS_8
-	str	w1, [x0, #UART_LCR]
-
-	/* Enable FIFOs, and clear receive and transmit */
-	mov	w1, #(UART_FCR_FIFO_EN | UART_FCR_CLEAR_RCVR |	\
-			UART_FCR_CLEAR_XMIT)
-	str	w1, [x0, #UART_FCR]
-
 	mov	w0, #1
-	ret
-core_init_fail:
-	mov	w0, wzr
 	ret
 endfunc console_core_init
