+++ plat/mediatek/mt6795/platform.mk
@@ -33,8 +33,8 @@ MTK_PLAT_SOC		:=	${MTK_PLAT}/${PLAT}
 
 # Add OEM customized codes
 OEMS				:= true
-# Mediatek RAM log
-MTK_RAM_LOG := 0
+MTK_SIP_KERNEL_BOOT_ENABLE := 1
+
 
 ifneq (${OEMS},none)
   OEMS_INCLUDES		:= -I${MTK_PLAT}/common/custom/
@@ -43,10 +43,6 @@ endif
 
 PLAT_INCLUDES		:=	-I${MTK_PLAT}/common/				\
 				-I${MTK_PLAT_SOC}/				\
-				-I${MTK_PLAT_SOC}/drivers/gpio/			\
-				-I${MTK_PLAT_SOC}/drivers/mtcmos/		\
-				-I${MTK_PLAT_SOC}/drivers/pmic/			\
-				-I${MTK_PLAT_SOC}/drivers/rtc/			\
 				-I${MTK_PLAT_SOC}/drivers/timer/		\
 				-I${MTK_PLAT_SOC}/drivers/uart/			\
 				-I${MTK_PLAT_SOC}/include/					\
@@ -59,36 +55,28 @@ PLAT_BL_COMMON_SOURCES	:=	lib/aarch64/xlat_tables.c			\
 				plat/common/plat_gic.c
 
 BL31_SOURCES		+=	drivers/arm/cci/cci.c				\
-				drivers/arm/gic/arm_gic.c			\
-				drivers/arm/gic/gic_v2.c \
-				drivers/arm/gic/gic_v3.c \
+				drivers/arm/gic/common/gic_common.c		\
+				drivers/arm/gic/v2/gicv2_main.c			\
+				drivers/arm/gic/v2/gicv2_helpers.c		\
+				plat/common/plat_gicv2.c			\
 				drivers/console/console.S			\
 				drivers/delay_timer/delay_timer.c		\
-				lib/cpus/aarch64/aem_generic.S			\
 				lib/cpus/aarch64/cortex_a53.S			\
-				lib/cpus/aarch64/cortex_a57.S			\
-				lib/cpus/aarch64/cortex_a72.S			\
 				plat/common/aarch64/platform_mp_stack.S		\
-        ${MTK_PLAT_SOC}/bl31_plat_setup.c		\
+				${MTK_PLAT_SOC}/bl31_plat_setup.c		\
+				${MTK_PLAT_SOC}/plat_mt_gic.c			\
 				${MTK_PLAT}/common/mtk_sip_svc.c		\
+				${MTK_PLAT}/common/platform_common.c		\
 				${MTK_PLAT_SOC}/aarch64/plat_helpers.S		\
-				${MTK_PLAT_SOC}/aarch64/platform_common.c	\
-				${MTK_PLAT_SOC}/drivers/pmic/pmic_wrap_init.c	\
-				${MTK_PLAT_SOC}/drivers/rtc/rtc.c		\
 				${MTK_PLAT_SOC}/drivers/timer/mt_cpuxgpt.c	\
 				${MTK_PLAT_SOC}/drivers/uart/8250_console.S	\
 				${MTK_PLAT_SOC}/plat_sip_svc.c	\
 				${MTK_PLAT_SOC}/plat_delay_timer.c		\
 				${MTK_PLAT_SOC}/plat_pm.c			\
-				${MTK_PLAT_SOC}/plat_sip_calls.c		\
 				${MTK_PLAT_SOC}/plat_topology.c			\
 				${MTK_PLAT_SOC}/power_tracer.c			\
 				${MTK_PLAT_SOC}/scu.c		\
 				${OEMS_SOURCES}
-
-ifeq (${MTK_RAM_LOG},1)
-BL31_SOURCES		+= ${MTK_PLAT}/common/log.c
-endif
 
 # Flag used by the MTK_platform port to determine the version of ARM GIC
 # architecture to use for interrupt management in EL3.
@@ -102,4 +90,5 @@ ERRATA_A53_836870	:=	1
 # indicate the reset vector address can be programmed
 PROGRAMMABLE_RESET_ADDRESS	:=	1
 
-$(eval $(call add_define,MTK_RAM_LOG))
+$(eval $(call add_define,MTK_SIP_KERNEL_BOOT_ENABLE))
+
