+++ plat/mediatek/mt8173/drivers/spm/spm.c
@@ -30,7 +30,7 @@
 #include <bakery_lock.h>
 #include <debug.h>
 #include <mmio.h>
-#include <mt8173_def.h>
+#include <platform_def.h>
 #include <spm.h>
 #include <spm_suspend.h>
